requirements.txt
open-clip-torch
streamlit
chromadb
pillow
numpy
tqdm